Noise generator, hooking up more buttons, math
There is now a Python 2.7 program which is able to generate noise for
the program so that it may be tested. It provides a CLI/TUI That can be
used to arm, disarm, takeoff, and land a UAV

Config was updated, now doesn't contain much only allowing for the
changing of the flight time. Other options were removed because they
have been standardized (etc., the directory to store results should be
in the C:\Users\Andrew\Documents folder)

Started to make a Session object to handle multiple flights

Hooked up more buttons, such as the continuing flight button

Calculations are not coming along, but the program does have a space to
handle the right/left percentages